This is hard because hound config won't go away
I wanted to allow configs to be initialized with content instead of
hound configs, so they could do a little less (SRP), and hound config
would be less of a god object (it's not actually that big but it's
passed around a ton for non-obvious reasons, which really irks me).
However, during the PR process I refined the interface I was shooting
for, and now (even though I still am irked) it doesn't matter if hound
config is still passed in to initialize configs, so I'm going to make
the simpler change instead.