WIP Expand linter shared examples
Adds specs for `#file_review` to the linter shared examples to help
clean up and make sure the linters are tested consistently. Downside is
all the linter specs load rails_helper and there's a few that have more
test coverage than the others. Some of that could probably be removed or
added to the shares examples, but this diff is big enough as-is so I'm
stopping here.

Needs #1338 and #1339 to be merged finishing up.